fix team settings mix of client/server side logic
